+++ src/Module/Acf.php
@@ -1,99 +1,71 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Sober\Controller\Module;
 
 use Sober\Controller\Utils;
 
 class Acf
 {
-    // Config
-    protected $data = [];
+    protected array $data = [];
 
-    private $returnArrayFormat = false;
+    private bool $returnArrayFormat = false;
 
-    /**
-     * Construct
-     *
-     * Initialise the Loader methods
-     */
     public function __construct()
     {
         $this->setReturnFilter();
     }
 
-    /**
-     * Set Return Filter
-     *
-     * Return filter sober/controller/acf-array
-     */
-    private function setReturnFilter()
+    private function setReturnFilter(): void
     {
-        $this->returnArrayFormat =
-            (has_filter('sober/controller/acf/array')
-            ? apply_filters('sober/controller/acf/array', $this->returnArrayFormat)
-            : false);
+        $this->returnArrayFormat = has_filter('sober/controller/acf/array')
+            ? (bool) apply_filters('sober/controller/acf/array', $this->returnArrayFormat)
+            : false;
     }
 
-    /**
-     * Iterates over array and adds a new snake cased key, with orignial value, for each kebab cased key
-     *
-     * Return void
-     */
-    private function recursiveSnakeCase(&$data) {
-      if(!is_array($data))
-        return;
-        
-      foreach ($data as $key => $val) {
-        if (is_array($val)) {
-          $this->recursiveSnakeCase($val);
-        } else {
-          $data[Utils::convertKebabCaseToSnakeCase($key)] = $val;
+    private function recursiveSnakeCase(array &$data): void
+    {
+        foreach ($data as $key => $val) {
+            $snakeKey = Utils::convertKebabCaseToSnakeCase($key);
+
+            if ($snakeKey !== $key) {
+                $data[$snakeKey] = $val;
+            }
+
+            if (is_array($val)) {
+                $this->recursiveSnakeCase($val);
+            }
         }
-      }
     }
 
-    /**
-     * Set Data Return Format
-     *
-     * Return object from array if acf/array filter is not set to true
-     */
-    public function setDataReturnFormat()
+    public function setDataReturnFormat(): void
     {
         if ($this->returnArrayFormat) {
             return;
         }
 
-        if ($this->data) {
-            foreach ($this->data as $key => $item) {
-                $this->data[$key] = json_decode(json_encode($item));
+        foreach ($this->data as $key => $item) {
+            if (is_array($item) || is_object($item)) {
+                $this->data[$key] = json_decode(json_encode($item), false);
             }
         }
     }
 
-    /**
-     * Set Data Options Page
-     *
-     * Set data from the options page
-     */
-    public function setDataOptionsPage()
+    public function setDataOptionsPage(): void
     {
         if (!function_exists('acf_add_options_page')) {
-            return [];
+            return;
         }
 
-        if (get_fields('options')) {
-            $this->data['acf_options'] = get_fields('options');
-        } else {
-            return [];
+        $fields = get_fields('options');
+
+        if (!empty($fields)) {
+            $this->data['acf_options'] = $fields;
         }
     }
 
-    /**
-     * Set Data
-     *
-     * Set data from passed in field keys
-     */
-    public function setData($acf)
+    public function setData(bool|string|array $acf): void
     {
         $query = get_queried_object();
 
@@ -101,31 +73,34 @@ class Acf
             return;
         }
 
-        if (is_bool($acf)) {
-            $this->data = get_fields($query);
+        if (is_bool($acf) && $acf) {
+            $fields = get_fields($query);
+            if (is_array($fields)) {
+                $this->data = $fields;
+            }
         }
 
         if (is_string($acf)) {
-            $this->data = [$acf => get_field($acf, $query)];
+            $value = get_field($acf, $query);
+            if ($value !== null) {
+                $this->data[$acf] = $value;
+            }
         }
 
         if (is_array($acf)) {
             foreach ($acf as $item) {
-                $this->data[$item] = get_field($item, $query);
+                $value = get_field($item, $query);
+                if ($value !== null) {
+                    $this->data[$item] = $value;
+                }
             }
         }
 
         $this->recursiveSnakeCase($this->data);
     }
 
-    /**
-     * Get Data
-     *
-     * Return the data
-     * @return array
-     */
-    public function getData()
+    public function getData(): array
     {
-        return is_array($this->data) ? $this->data : [];
+        return $this->data;
     }
 }