+++ src/Module/Tree.php
@@ -1,7 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Sober\Controller\Module;
 
+/**
+ * Marker interface for Controllers that should inherit data
+ * from less-specific Controllers in the template hierarchy.
+ */
 interface Tree
 {
 }