+++ src/Utils.php
@@ -1,83 +1,44 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Sober\Controller;
 
 class Utils
 {
-    /**
-     * Is File PHP
-     *
-     * Determine if the file is a PHP file
-     * @return boolean
-     */
-    public static function isFilePhp($filename)
+    public static function isFilePhp(string $filename): bool
     {
-        return (in_array(pathinfo($filename, PATHINFO_EXTENSION), ['php']));
+        return pathinfo($filename, PATHINFO_EXTENSION) === 'php';
     }
 
-    /**
-     * Does File Contain
-     *
-     * Determine if the file contains a string
-     * @return boolean
-     */
-    public static function doesFileContain($filename, $str)
+    public static function doesFileContain(string $filename, string $str): bool
     {
-        return strpos(file_get_contents($filename), $str) !== false;
+        $contents = @file_get_contents($filename);
+        return $contents !== false && str_contains($contents, $str);
     }
 
-    /**
-     * Is Array Indexed
-     *
-     * Determine if the array is indexed
-     * @return boolean
-     */
-    public static function isArrayIndexed(array $array)
+    public static function isArrayIndexed(array $array): bool
     {
         return array_keys($array) === range(0, count($array) - 1);
     }
 
-    /**
-     * Does String Contain Markup
-     *
-     * Determine if the string contains markup
-     * @return boolean
-     */
-    public static function doesStringContainMarkup($str)
+    public static function doesStringContainMarkup(string $str): bool
     {
-        return (is_string($str) && $str !== strip_tags($str));
+        return $str !== strip_tags($str);
     }
 
-    /**
-     * Convert To Snake Case
-     *
-     * Convert camel case to snake case for data variables
-     * @return string
-     */
-    public static function convertToSnakeCase($str)
+    public static function convertToSnakeCase(string $str): string
     {
         return strtolower(preg_replace('/(?<!^)[A-Z]/', '_$0', $str));
     }
 
-    /**
-     * Convert To Kebab Case
-     *
-     * Convert camel case to kebab case for templates
-     * @return string
-     */
-    public static function convertToKebabCase($str)
+    public static function convertToKebabCase(string $str): string
     {
         return strtolower(preg_replace('/(?<!^)[A-Z]/', '-$0', $str));
     }
 
-    /**
-     * Convert To Snake Case
-     *
-     * Converts kebab case to snake case for data variables
-     * @return string
-     */
-    public static function convertKebabCaseToSnakeCase($str)
+    public static function convertKebabCaseToSnakeCase(string $str): string
     {
-      return strtolower(str_replace('-', '_', $str));
+        return strtolower(str_replace('-', '_', $str));
     }
 }